feat: add pages slug
