+++ app/pages/conference/[...slug].vue
@@ -1,0 +1,9 @@
+<script setup lang="ts">
+definePageMeta({
+  layout: 'conference',
+})
+</script>
+
+<template>
+  <ContentView root="conference" />
+</template>
