+++ app/pages/events/[...slug].vue
@@ -1,0 +1,9 @@
+<script setup lang="ts">
+definePageMeta({
+  layout: 'conference',
+})
+</script>
+
+<template>
+  <ContentView root="events" />
+</template>
